+++ source/TUtils.Common.EF6/Transaction/ITransactionService.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Data.Entity;
+using TUtils.Common.Transaction;
 
 namespace TUtils.Common.EF6.Transaction
 {
@@ -7,17 +8,9 @@ namespace TUtils.Common.EF6.Transaction
 	/// provides DbContext and encapsulates actions with a transaction.
 	/// </summary>
 	/// <typeparam name="TDbContext"></typeparam>
-	public interface ITransactionService<TDbContext>
+	public interface ITransactionService<TDbContext> : ITransactionService
 		where TDbContext : DbContext
 	{
-		/// <summary>
-		/// all transactions called in "action" will use the same DbContext object if possible.
-		/// May be nested
-		/// </summary>
-		/// <param name="action">
-		/// 
-		/// </param>
-		void DoWithSameDbContext(Action action);
 
 		/// <summary>
 		/// May be nested
@@ -32,5 +25,19 @@ namespace TUtils.Common.EF6.Transaction
 		/// called for example in case of a lost update situation
 		/// </param>
 		void DoInTransaction(Action<TDbContext> action, Action onConcurrencyException);
+
+		/// <summary>
+		/// May be nested
+		/// </summary>
+		/// <param name="action"></param>
+		T DoInTransaction<T>(Func<TDbContext,T> action);
+		/// <summary>
+		/// May be nested
+		/// </summary>
+		/// <param name="action"></param>
+		/// <param name="onConcurrencyException">
+		/// called for example in case of a lost update situation
+		/// </param>
+		T DoInTransaction<T>(Func<TDbContext,T> action, Action onConcurrencyException);
 	}
 }