+++ source/Testprogramm/Student.cs
@@ -1,9 +1,0 @@
-﻿namespace Testprogramm
-{
-	public class Student
-	{
-		public int Id { get; set; }
-		public  string Name { get; set; }
-
-	}
-}