+++ documentation/history.md
@@ -1,5 +1,24 @@
 # History of TUtils.Common
 
+## Version 1.0.1.4 - 20.10.2016
+new features added: 
+	- TUtils.Common: Debouncer, IDebouncer, Debouncer, ILazy, Lazy, ITransactionService
+	- TUtils.Common.EF6: 
+		- TransactionService now implements new interface ITransactionService, too.
+		- EntityExtension.ApplyChanges()
+	- TUtils.Common.MVC: 
+		- Html.RenderConfirmationModalDlg()
+		- Html.RenderMailLink()
+		- Html.RenderConfirmationModalDlg()
+		- string.UrlEncoded()
+		- string.UrlDecode()
+		- byte[].ToUrlEncodedString()
+		- byte[].ToBytesFromUrlEncodedString()
+		- Uri.GetQueryParameter()
+		- Uri.QueryParamters()
+		- Uri.RemoveQuery()
+		- Uri.AddQueryParameters()
+		- Uri.RemoveQueryParameter()
 ## Version 1.0.1.3 - 17.08.2016
 new features added: IDIContainer, TUtils.Common.MVC.dll, 
 ## Version 1.0.1.2 - 26.07.2016